+++ src/pages/teacherDashboard/classes/CreateClassForm.tsx
@@ -1,18 +1,12 @@
-import * as forms from "codeforlife/components/form"
-import { Stack, Typography } from "@mui/material"
 import { type FC } from "react"
 import { type SchoolTeacherUser } from "codeforlife/api"
+import { Typography } from "@mui/material"
 import { generatePath } from "react-router"
 import { useNavigate } from "codeforlife/hooks"
 
-import {
-  ClassNameField,
-  ReadClassmatesDataField,
-  TeacherAutocompleteField,
-} from "../../../components/form"
+import { CreateClassForm as Form } from "../../../components/form"
 import { type RetrieveUserResult } from "../../../api/user"
 import { paths } from "../../../routes"
-import { useCreateClassMutation } from "../../../api/klass"
 
 export interface CreateClassFormProps {
   authUser: SchoolTeacherUser<RetrieveUserResult>
@@ -26,18 +20,8 @@ const CreateClassForm: FC<CreateClassFormProps> = ({ authUser }) => {
       <Typography variant="h5" sx={{ mb: 2 }}>
         Create a new class
       </Typography>
-      <Typography>
-        {authUser.teacher.is_admin
-          ? "When you set up a new class, a unique class access code will automatically be generated for the teacher assigned to the class."
-          : "When you set up a new class, a unique class access code will automatically be generated, with you being identified as the teacher for that class."}
-      </Typography>
-      <forms.Form
-        initialValues={{
-          name: "",
-          teacher: authUser.teacher.id,
-          read_classmates_data: false,
-        }}
-        useMutation={useCreateClassMutation}
+      <Form
+        authUser={authUser}
         submitOptions={{
           then: ({ id }, { name }) => {
             navigate(
@@ -58,16 +42,7 @@ const CreateClassForm: FC<CreateClassFormProps> = ({ authUser }) => {
             )
           },
         }}
-      >
-        <Stack gap={2}>
-          <Stack direction={{ sm: "row" }} gap={2}>
-            <ClassNameField required />
-            {authUser.teacher.is_admin && <TeacherAutocompleteField required />}
-          </Stack>
-          <ReadClassmatesDataField />
-          <forms.SubmitButton>Create class</forms.SubmitButton>
-        </Stack>
-      </forms.Form>
+      />
     </>
   )
 }